# documentation/api.yaml
openapi: 3.0.3
info:
  title: Westco Tack API Course Project Documentation.
  version: 1.0.11
  license:
    name: Apache 2.0
    url: "http://apache.org/"
servers:
  - url: https://educations-apis.herokuapp.com/
    description: Https Hosting Heroku
  - url: http://localhost:5000
    description: Dev Server on Port
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /superAdmin/signUp:
    post:
      summary: Sign Up Super Admin
      tags:
        - Register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                secretInfo:
                  type: string
              example:
                name: Sunna
                email: Sunna@gmail.com
                password: "12345Sunna"
                secretInfo: Men Uzbekistonda yashayman!
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id: integer
                  name: string
                  email: string
                  password: string
                  secretInfo: string
                  createdAt: string
                  updatedAt: string
              example:
                _id: 62af3d203816e616bd006514
                name: Sunnatullox
                email: Sunnatullox@gmail.com
                password: $2a$10$vF7DFPtLrXo4FXpO2vm7U.GXPHpmJWgVWD51rVsXVgL84cAzjyPrW
                secretInfo: $2a$10$p/oRKgRopkRvx5/a/MTp/OEnhVk7wx9i8SIKJkVFPPipQGUUBSX2S
                createdAt: 2022-06-19T15:13:36.493+00:00
                updatedAt: 2022-06-20T14:01:43.788+00:00
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /superAdmin/signIn:
    post:
      summary: Sign In Super Admin
      tags:
        - Register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
              example:
                email: Sunna@gmail.com
                password: "12345Sunna"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: integer
                  _id: integer
                  name: string
                  email: string
                  password: string
                  secretInfo: string
                  createdAt: string
                  updatedAt: string
              example:
                token: TU3ODExNjB9.RoCiHr5wflJbYUWs67NIK4IL3PQcM6fTDg2eeBGsJWo
                _id: 62af3d203816e616bd006514
                name: Sunnatullox
                email: Sunnatullox@gmail.com
                password: $2a$10$vF7DFPtLrXo4FXpO2vm7U.GXPHpmJWgVWD51rVsXVgL84cAzjyPrW
                secretInfo: $2a$10$p/oRKgRopkRvx5/a/MTp/OEnhVk7wx9i8SIKJkVFPPipQGUUBSX2S
                createdAt: 2022-06-19T15:13:36.493+00:00
                updatedAt: 2022-06-20T14:01:43.788+00:00
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /superAdmin/createCategry:
    post:
      summary: Sign In Super Admin
      tags:
        - Super Admin
      description: Super Admin created Category
      consumes:
        - multipart/form-data
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                keywords:
                    type: array
                categoryImage:
                  type: string
                  format: binary
                categoryId:
                  type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id: 
                    type: string
                  title:
                    type: string
                  description: 
                    type: string
                  categoryImage:
                    type: string
                  keywords: 
                    type: array
                  isActive: 
                    type: string
                  subcatigory:
                    type: array
                  createdAt: 
                    type: string
                  updatedAt: 
                    type: string
              example:
                title: JavaScript
                description: JavaScript category
                categoryImage: /196544b0-7690-4348-a96d-4daf06ef1a08C.jpg
                keywords: []
                isActive: true
                _id: 636d258bf85413dea3d454d2
                subcatigory: []
                createdAt: 2022-11-10T16:23:39.238Z
                updatedAt: 2022-11-10T16:23:39.238Z
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /superAdmin/getCategrys:
    get:
      summary: Get one course user
      tags:
        - Super Admin
      description: Get one course comment
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id: 
                    type: string
                  name: 
                    type: string
                  courseId:
                    type: string
                  userId: 
                    type: string
                  description:
                    type: string
                  createdAt:
                    type: string
                  updatedAt:
                    type: string
              example:
                "_id": "62ae9ca688144d2d246e879d"
                "courseId": "62ae95af49ef0d92e405b231"
                "name": "Sunna"
                "description": "Assalamu Aleykum menga bu React Next JS darslari juda maqul bo'ldi "
                "userId": "62ac7f3fa525215eeede092f"
                "createdAt": "2022-06-19T03:48:54.490Z"
                "updatedAt": "2022-06-19T03:48:54.490Z"
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /superAdmin/updateCategry:
    put:
      summary: Sign In Super Admin
      tags:
        - Super Admin
      description: Super Admin update Category
      consumes:
        - multipart/form-data
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                keywords:
                    type: array
                isActive:
                  type: boolean
                categoryImage:
                  type: string
                  format: binary
                categoryId:
                  type: string
                subCategoryId:
                  type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id: 
                    type: string
                  title:
                    type: string
                  description: 
                    type: string
                  categoryImage:
                    type: string
                  keywords: 
                    type: array
                  isActive: 
                    type: string
                  subcatigory:
                    type: array
                  createdAt: 
                    type: string
                  updatedAt: 
                    type: string
              example:
                title: JavaScript
                description: JavaScript category
                categoryImage: /196544b0-7690-4348-a96d-4daf06ef1a08C.jpg
                keywords: []
                isActive: true
                _id: 636d258bf85413dea3d454d2
                subcatigory: []
                createdAt: 2022-11-10T16:23:39.238Z
                updatedAt: 2022-11-10T16:23:39.238Z
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /superAdmin/deletedCategry:
    delete:
      summary: Sign In Super Admin
      tags:
        - Super Admin
      description: Super Admin created Category
      consumes:
        - multipart/form-data
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                categoryId:
                  type: string
                subCategoryId:
                  type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: 
                    type: string
              example:
                message: category yokiy subCategory muvafaqiyatli o'chirildi.
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /superAdmin/updateIsActiveCourse:
    put:
      summary: Sign In Super Admin
      tags:
        - Super Admin
      description: Super Admin created Category
      consumes:
        - multipart/form-data
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                isActive:
                  type: boolean
                courseId:
                  type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: 
                    type: string
              example:
                message: Course muvafaqiyatli active okiy neActive bo'ldi.
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /superAdmin/updateIsActiveLesson:
    put:
      summary: Sign In Super Admin
      tags:
        - Super Admin
      description: Super Admin created Category
      consumes:
        - multipart/form-data
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                isActive:
                  type: boolean
                lessonId:
                  type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: 
                    type: string
              example:
                message: lesson(dars) muvafaqiyatli active okiy neActive bo'ldi.
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /superAdmin/updateIsActiveCatigory:
    put:
      summary: Sign In Super Admin
      tags:
        - Super Admin
      description: Super Admin created Category
      consumes:
        - multipart/form-data
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                isActive:
                  type: boolean
                categoryId:
                  type: string
                subCategoryId:
                  type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: 
                    type: string
              example:
                message: category yokiy subCategory muvafaqiyatli active okiy neActive bo'ldi.
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /superAdmin/usersStatus:
    get:
      summary: Get All Users registered platform status
      tags:
        - Super Admin
      description: Get All Users registered Platform
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
              example:
                users:
                  _id: 1
                  total: 19
        404:
          description: Not Found user status
        500:
          description: Internal Server Error
...